OMAGOSH THE STINKING THING FOR THE ARM POSITIONS WORK WOAHHHH
What it says on the tin. Also had to fix some of the naming conventions and whatever you set the claw names to and armmotor had to reversed. There's a gobuilda so the positons had to be changed to account for its new torque, also auto has to be changed now beause the claw is like completely different.